Add ESP reset via GPIO on load and unload
Adds a parameter to toggle the CH_PD reset on the ESP, which should
be hooked to a GPIO addressable via sysfs.  The reset happens on
module load in case no previous reset has been asserted (as on cold boot)
and on unload so that the ESP properly identifies on the SD bus on
future warm boots (so it's auto loaded as a module dependency).

A new parameter (esp_reset_gpio) can be used to set any kernel GPIO driver
mapped GPIO as the reset pin, or the default of 0 can be used.  On the
Raspberry Pi, GPIO 0 is the ID_SD signal.

Also fixed a build warning and a mac80211 function parameter mismatch.  The
driver now builds cleanly.